+++ packages/elements/mention/examples/08-resourced-mention-on-n20-background.tsx
@@ -3,10 +3,8 @@ import { AnalyticsListener } from '@atlaskit/analytics';
 import { colors } from '@atlaskit/theme';
 import debug from '../src/util/logger';
 import { onMentionEvent } from '../example-helpers/index';
-import {
-  mockMentionData as mentionData,
-  mockMentionProvider as mentionProvider,
-} from '../src/__tests__/unit/_test-helpers';
+import { mockMentionData as mentionData } from '../src/__tests__/unit/_test-helpers';
+import { MockMentionResource } from '@atlaskit/util-data-test';
 import ResourcedMention from '../src/components/Mention/ResourcedMention';
 import { IntlProvider } from 'react-intl';
 
@@ -23,6 +21,8 @@ function listenerHandler(eventName: string, eventData: Object) {
 }
 
 export default function Example() {
+  const mentionProvider = Promise.resolve(new MockMentionResource({}));
+
   return (
     <IntlProvider locale="en">
       <div style={style}>
@@ -52,6 +52,16 @@ export default function Example() {
           <ResourcedMention
             {...mentionData}
             accessLevel={'NONE'}
+            mentionProvider={mentionProvider}
+            onClick={onMentionEvent}
+            onMouseEnter={onMentionEvent}
+            onMouseLeave={onMentionEvent}
+          />
+        </div>
+        <div style={padding}>
+          <ResourcedMention
+            {...mentionData}
+            text=""
             mentionProvider={mentionProvider}
             onClick={onMentionEvent}
             onMouseEnter={onMentionEvent}