+++ packages/elements/mention/src/index.ts
@@ -2,14 +2,21 @@ import MentionResource, {
   AbstractMentionResource,
   MentionContextIdentifier,
   MentionProvider,
+  ResolvingMentionProvider,
   MentionStats,
   MentionResourceConfig,
+  isResolvingMentionProvider,
 } from './api/MentionResource';
 import TeamMentionResource from './api/TeamMentionResource';
 import PresenceResource, {
   PresenceProvider,
   AbstractPresenceResource,
 } from './api/PresenceResource';
+import {
+  DefaultMentionNameResolver,
+  MentionNameResolver,
+} from './api/MentionNameResolver';
+import { MentionNameClient } from './api/MentionNameClient';
 import MentionItem from './components/MentionItem';
 import MentionList from './components/MentionList';
 import ResourcedMentionList from './components/ResourcedMentionList';
@@ -19,6 +26,8 @@ import ResourcedMention from './components/Mention/ResourcedMention';
 import {
   MentionDescription,
   MentionsResult,
+  MentionNameStatus,
+  MentionNameDetails,
   isSpecialMention,
   TeamMember,
 } from './types';
@@ -31,13 +40,19 @@ export {
   MentionResource,
   TeamMentionResource,
   PresenceResource,
+  DefaultMentionNameResolver,
   AbstractMentionResource,
   AbstractPresenceResource,
   // Interfaces
+  ResolvingMentionProvider,
   MentionProvider,
   PresenceProvider,
   MentionDescription,
   MentionsResult,
+  MentionNameResolver,
+  MentionNameClient,
+  MentionNameStatus,
+  MentionNameDetails,
   // types
   MentionContextIdentifier,
   MentionStats,
@@ -52,6 +67,7 @@ export {
   ResourcedMention,
   // Functions
   isSpecialMention,
+  isResolvingMentionProvider,
   // Constants
   ELEMENTS_CHANNEL,
 };