+++ packages/search/global-search/src/components/jira/JiraQuickSearchContainer.tsx
@@ -232,7 +232,7 @@ export class JiraQuickSearchContainer extends React.Component<
     recentItems,
     keepPreQueryState,
     searchSessionId,
-  }: SearchResultProps) => {
+  }: SearchResultProps<GenericResultMap>) => {
     const query = latestSearchQuery;
     const {
       referralContextIdentifiers,
@@ -395,11 +395,16 @@ export class JiraQuickSearchContainer extends React.Component<
         [],
         getJiraMaxObjects(features.abTest, JIRA_PREQUERY_RESULT_LIMIT),
       )
-      .then(xpRecentResults => ({
-        objects: xpRecentResults.results.get(Scope.JiraIssue) || [],
-        containers:
-          xpRecentResults.results.get(Scope.JiraBoardProjectFilter) || [],
-      }));
+      .then(xpRecentResults => {
+        const objects = xpRecentResults.results[Scope.JiraIssue];
+        const containers =
+          xpRecentResults.results[Scope.JiraBoardProjectFilter];
+
+        return {
+          objects: objects ? objects.items : [],
+          containers: containers ? containers.items : [],
+        };
+      });
   };
 
   getJiraRecentItems = (sessionId: string): Promise<GenericResultMap> => {
@@ -443,7 +448,9 @@ export class JiraQuickSearchContainer extends React.Component<
     }
   };
 
-  getRecentItems = (sessionId: string): Promise<ResultsWithTiming> => {
+  getRecentItems = (
+    sessionId: string,
+  ): Promise<ResultsWithTiming<GenericResultMap>> => {
     return Promise.all([
       this.getJiraRecentItems(sessionId),
       this.getRecentlyInteractedPeople(),
@@ -452,7 +459,7 @@ export class JiraQuickSearchContainer extends React.Component<
       .then(([jiraItems, people, canSearchUsers]) => {
         return { ...jiraItems, people: canSearchUsers ? people : [] };
       })
-      .then(results => ({ results } as ResultsWithTiming));
+      .then(results => ({ results } as ResultsWithTiming<GenericResultMap>));
   };
 
   getSearchResults = (
@@ -460,7 +467,7 @@ export class JiraQuickSearchContainer extends React.Component<
     sessionId: string,
     startTime: number,
     queryVersion: number,
-  ): Promise<ResultsWithTiming> => {
+  ): Promise<ResultsWithTiming<GenericResultMap>> => {
     const { features } = this.props;
 
     const crossProductSearchPromise = this.props.crossProductSearchClient.search(
@@ -500,20 +507,25 @@ export class JiraQuickSearchContainer extends React.Component<
         peopleElapsedMs,
         canSearchPeople,
       ]) => {
-        this.highlightMatchingFirstResult(query, xpsearchResults.results.get(
-          Scope.JiraIssue,
-        ) as JiraResult[]);
+        const objects = xpsearchResults.results[Scope.JiraIssue];
+        const containers =
+          xpsearchResults.results[Scope.JiraBoardProjectFilter];
+
+        const objectItems = objects ? objects.items : [];
+
+        this.highlightMatchingFirstResult(query, objectItems as JiraResult[]);
+
         return {
           results: {
-            objects: xpsearchResults.results.get(Scope.JiraIssue) || [],
-            containers:
-              xpsearchResults.results.get(Scope.JiraBoardProjectFilter) || [],
+            objects: objectItems,
+            containers: containers ? containers.items : [],
             people: canSearchPeople ? peopleResults : [],
           },
           timings: {
             crossProductSearchElapsedMs,
             peopleElapsedMs,
           },
+
           abTest: xpsearchResults.abTest,
         };
       },